fix bug in _remove_unsupported_cxx_versions_for_gcc()
- moved unittest in extra file
