+++ .github/workflows/testDeploy.yml
@@ -18,7 +18,7 @@ jobs:
               run: |
                 pip install .
             - name: run unit tests
-              run: python -m unittest discover -s tests
+              run: python -m unittest discover -s tests -p "*.py"
     integration-tests:
       needs: unit-tests
       runs-on: ubuntu-latest
